add possibility to add new board manually
Until recently, new boards were only created on server startup.
However, that was not really optimal for the players and was just
a workaround until the application could offer something better.
This workaround has been disabled in one of the previous commits,
so the player can now decide when to create a new board.