point "back" button in settings to board
It was pointing to board selection, but that is counter-intuitive
and just requires one extra user interaction to get back to the
board. Since the settings can currently only be reached directly
from the board anyway, it is the natural thing to do to go back
to the board, once the settings have been adjiusted.