adjust Dockerfile so that MongoDB starts as expected
MongoDB expects some locale-related variables to be set properly
on the system. It will not start, if these are not set.
Basically, LC_ALL needs to be set and the proper locale must
have been generated beforehand. The container will use the
en_GB.UTF-8 locale for the moment.

See <https://github.com/meteor/meteor/issues/4019> for more
information about that issue.

Additionally, a short version history has been added to the
Dockerfile.